+++ src/comp3415/telehealth/db/logInfo.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package comp3415.telehealth;
+package comp3415.telehealth.db;
 
 /**
  *
