+++ src/comp3415/telehealth/db/logout.java
@@ -3,7 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package comp3415.telehealth;
+package comp3415.telehealth.db;
+import comp3415.telehealth.views.loginLayout;
+
 import javax.swing.JFrame;
 /**
  *
