+++ src/comp3415/telehealth/views/login.fxml
@@ -6,7 +6,6 @@
 
 <GridPane fx:controller="comp3415.telehealth.LoginController" alignment="CENTER" gridLinesVisible="true" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
    <children>
-
        <!-- Bottom Toolbar -->
       <ToolBar fx:id="bottomBar" prefHeight="40.0" GridPane.hgrow="NEVER" GridPane.rowIndex="1" GridPane.valignment="BOTTOM" GridPane.vgrow="NEVER">
         <items>
@@ -26,7 +25,6 @@
           </AnchorPane>
         </content>
       </TitledPane>
-
    </children>
    <padding>
       <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
