+++ src/comp3415/telehealth/views/loginLayout.java
@@ -3,7 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package comp3415.telehealth;
+package comp3415.telehealth.views;
+
+import comp3415.telehealth.db.login;
 
 import javax.swing.JOptionPane;
